voglio solo vedere il codice sul portatile

+++ algorithm.h
@@ -4,7 +4,7 @@
 
 #ifndef ALGORITHM_PROJECT_2019_ALGORITHM_H
 #define ALGORITHM_PROJECT_2019_ALGORITHM_H
-void start_algorithm_print(double *array, int size);
-void start_algorithm(double *array, int size);
+void start_algorithm_print(double *array, long size);
+void start_algorithm(double *array, long size);
 
 #endif //ALGORITHM_PROJECT_2019_ALGORITHM_H
